+++ 11-21H2/ntdll/LdrpPrepareModuleForExecution.c
@@ -25,7 +25,7 @@ __int64 __fastcall LdrpPrepareModuleForExecution(__int64 a1, __int64 a2)
   char v10; // [rsp+50h] [rbp+18h] BYREF
 
   v4 = 0;
-  if ( RtlIsCriticalSectionLockedByThread((__int64)&LdrpDllNotificationLock) )
+  if ( RtlIsCriticalSectionLockedByThread(&LdrpDllNotificationLock) )
     return v4;
   v5 = *(_QWORD *)(a1 + 152);
   switch ( *(_DWORD *)(v5 + 56) )
