+++ 11-22H2/ntdll/RtlPcToFileHeader.c
@@ -15,7 +15,7 @@ PVOID __stdcall RtlPcToFileHeader(PVOID PcValue, PVOID *BaseOfImage)
     || (unsigned __int64)PcValue >= *((_QWORD *)&xmmword_180199520 + 1)
                                   + (unsigned __int64)(unsigned int)qword_180199530 )
   {
-    RtlpxLookupFunctionTable((unsigned __int64)PcValue, (__int64 *)&v4);
+    RtlpxLookupFunctionTable(PcValue, (__int64 *)&v4);
   }
   else
   {
