+++ 11-22H2/ntoskrnl/CmpCopySaclToVirtualKey.c
@@ -107,11 +107,7 @@ __int64 __fastcall CmpCopySaclToVirtualKey(
   if ( SaclSecurityDescriptor >= 0 )
   {
     WORD1(SecurityDescriptor[0]) = v15[1] & 0x7FFF;
-    SaclSecurityDescriptor = RtlSetSaclSecurityDescriptor(
-                               (__int64)SecurityDescriptor,
-                               SaclPresent,
-                               (__int64)Sacl,
-                               SaclDefaulted);
+    SaclSecurityDescriptor = RtlSetSaclSecurityDescriptor(SecurityDescriptor, SaclPresent, Sacl, SaclDefaulted);
     if ( SaclSecurityDescriptor >= 0 )
     {
       SaclSecurityDescriptor = RtlGetOwnerSecurityDescriptor(v15, &Owner, &SaclDefaulted);
