+++ 11-22H2/ntoskrnl/IoReadPartitionTable.c
@@ -23,13 +23,13 @@ NTSTATUS __stdcall IoReadPartitionTable(
   struct _DEVICE_OBJECT *v7; // r10
   NTSTATUS v8; // edi
   int PartitionTable; // eax
-  ULONG *v10; // rbx
+  DWORD *v10; // rbx
   SIZE_T v11; // rbp
   struct _DRIVE_LAYOUT_INFORMATION *PoolWithTag; // rax
   __int64 v13; // r8
   struct _DRIVE_LAYOUT_INFORMATION *v14; // r10
   __int64 v15; // rcx
-  ULONG *v16; // r9
+  DWORD *v16; // r9
   char v17; // al
   PVOID P; // [rsp+20h] [rbp-1C8h] BYREF
   _QWORD v20[50]; // [rsp+30h] [rbp-1B8h] BYREF
@@ -43,7 +43,7 @@ NTSTATUS __stdcall IoReadPartitionTable(
   if ( v8 >= 0 )
   {
     PartitionTable = SC_DISK::ReadPartitionTable((SC_DISK *)v20, (struct SC_DISK_LAYOUT **)&P);
-    v10 = (ULONG *)P;
+    v10 = (DWORD *)P;
     v8 = PartitionTable;
     if ( PartitionTable >= 0 )
     {
